feat(ws-25): per-run role overrides (StartRun.role_overrides), end to end
WS-22 could validate a Kohral turn's `model` and record it on the ledger but
not apply it: StartRun had no per-run role field, and mutating [roles] in a
daemon that serves other runs concurrently is not an option. This is the plan
change plus its implementation.

- plan/02 §Run and the event catalog: StartRun and run.started carry
  `role_overrides: BTreeMap<String, ModelAlias>`, keyed by role name.
- plan/05 §3.1: what each key overrides, and the one case still open (the
  run-level judge goes through EvaluatorPort, which does not see the run).
- kevin-domain: RoleOverrides on the command, the event and the aggregate.
- kevin-orchestrator: role_route() prefers the override and names it in the
  error; a `default` override bypasses the router so every plan task runs on
  the requested alias; the phase jobs carry the overrides.
- kevin-kohral: a resolved turn model becomes the override for planner,
  clarifier, judge, integrator and default.

Also makes the timing-sensitive ac_ws08 scenarios robust under load without
weakening them: Handle::abort_and_join() waits for the killed actors instead of
sleeping 50ms, the wait deadline is 90s (KEVIN_TEST_WAIT_SECS), the drain
scenario holds its attempt rather than racing a 3s delay, and the fan-out
scenario's attempts outlast a scheduling round.

+++ crates/kevin-orchestrator/tests/ac_ws25_load.rs
@@ -34,7 +34,7 @@ use kevin_worker::fake::{Rule, Scenario};
 use tokio_util::sync::CancellationToken;
 
 /// How many plan tasks the run carries.
-const TASKS: usize = 50;
+const TASKS: u32 = 50;
 /// The bulkhead under test.
 const MAX_PARALLEL: u16 = 8;
 /// Generous: a fake attempt takes ~30 ms, so 50 of them at 8-wide is seconds.
@@ -92,7 +92,7 @@ async fn ac_ws25_10_1_fifty_fake_tasks_respect_the_bulkheads_and_stay_bounded() 
     let roles = Arc::new(
         ScriptedRoles::new()
             .with_understanding(understanding("fifty small things"))
-            .with_plan(plan_of(TASKS)),
+            .with_plan(plan_of(TASKS as usize)),
     );
     // Each attempt is slower than the saga tick, so attempts genuinely pile up
     // and the semaphore is exercised instead of the run trickling one task per
@@ -112,7 +112,7 @@ async fn ac_ws25_10_1_fifty_fake_tasks_respect_the_bulkheads_and_stay_bounded() 
                 config.budget.max_parallel_tasks = 64;
                 // The default plan cap is 24 tasks; this scenario is about
                 // what happens past it.
-                config.orchestrator.max_tasks_per_run = TASKS as u32;
+                config.orchestrator.max_tasks_per_run = TASKS;
             }),
     )
     .await;
@@ -169,12 +169,12 @@ async fn ac_ws25_10_1_fifty_fake_tasks_respect_the_bulkheads_and_stay_bounded() 
     // -- the run actually did the work ---------------------------------------
     assert_eq!(
         count(&events, "task.created"),
-        TASKS,
+        TASKS as usize,
         "the plan did not produce {TASKS} tasks"
     );
     assert_eq!(
         count(&events, "task.attempt_succeeded"),
-        TASKS,
+        TASKS as usize,
         "not every task succeeded"
     );
     assert_eq!(count(&events, "run.completed"), 1);
@@ -216,7 +216,7 @@ async fn ac_ws25_10_1_fifty_fake_tasks_respect_the_bulkheads_and_stay_bounded() 
         .fetch_one(harness.db.pool())
         .await
         .expect("count task_board");
-    assert_eq!(rows, TASKS as i64, "the board is missing tasks");
+    assert_eq!(rows, i64::from(TASKS), "the board is missing tasks");
 
     // -- memory stayed bounded ------------------------------------------------
     // The engine keeps per-*attempt* state, so 50 tasks at 8-wide must not
